remove load op
