+++ paddle/fluid/operators/load_op.cc
@@ -12,9 +12,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 
-#include "paddle/fluid/operators/load_op.h"
-
 #include <string>
+
+#include "paddle/fluid/framework/op_registry.h"
 
 namespace paddle {
 namespace operators {
@@ -65,12 +65,3 @@ class LoadOpProtoMaker : public framework::OpProtoAndCheckerMaker {
 namespace ops = paddle::operators;
 
 REGISTER_OPERATOR(load, ops::LoadOp, ops::LoadOpProtoMaker);
-
-REGISTER_OP_CPU_KERNEL(
-    load,
-    ops::LoadOpKernel<phi::CPUContext, float>,
-    ops::LoadOpKernel<phi::CPUContext, double>,
-    ops::LoadOpKernel<phi::CPUContext, paddle::platform::bfloat16>,
-    ops::LoadOpKernel<phi::CPUContext, int>,
-    ops::LoadOpKernel<phi::CPUContext, int8_t>,
-    ops::LoadOpKernel<phi::CPUContext, int64_t>);
