--- paddle/fluid/operators/load_op_xpu.cc
@@ -1,29 +1,0 @@
-/* Copyright (c) 2020 PaddlePaddle Authors. All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License. */
-
-#ifdef PADDLE_WITH_XPU
-
-#include "paddle/fluid/operators/load_op.h"
-
-namespace ops = paddle::operators;
-
-REGISTER_OP_XPU_KERNEL(
-    load,
-    ops::LoadOpKernel<paddle::platform::XPUDeviceContext, float>,
-    ops::LoadOpKernel<paddle::platform::XPUDeviceContext, double>,
-    ops::LoadOpKernel<paddle::platform::XPUDeviceContext, int>,
-    ops::LoadOpKernel<paddle::platform::XPUDeviceContext, int8_t>,
-    ops::LoadOpKernel<paddle::platform::XPUDeviceContext, int64_t>);
-
-#endif  // PADDLE_WITH_XPU
